api/requirements.txt
pydantic 
requests
fastapi
uvicorn
numpy
pillow
tensorflow-cpu
scikit-learn==1.0.2
scikit-image
python-multipart